[ca/yrup] Several fixes to ensure it works
Essentially I made sure that when an error occurs the function exits,
but not the shell (exit->return; () -> {}). Also adding in check for
mods/ra2 in yr, and making sure it copies it from $GHUBO/ra2 if it
doesn't exist. Removing old AppImages for the mod in question.

+++ obs/openra/ca.sh
@@ -2,36 +2,47 @@
 # Last commit with OBS version is b76be97a97f499f1b1b716418c36f95f8483d17a
 function caup {
 	printf "pushin' into $GHUBO/CAmod.\n"
-	pushd $GHUBO/CAmod || printf "pushdin' into $GHUBO/CAmod failed.\n" && exit
+	pushd $GHUBO/CAmod || {printf "pushdin' into $GHUBO/CAmod failed.\n" && return}
 	printf "Git pulling repository.\n"
-	git pull origin master -q || printf "Git pullin' repository failed.\n" && exit
+	git pull origin master -q || {printf "Git pullin' repository failed.\n" && return}
     # OpenRA latest engine version
-#    enlv=$(grep '^ENGINE\_VERSION' < mod.config | cut -d '"' -f 2)
+#    enlv=${grep '^ENGINE\_VERSION' < mod.config | cut -d '"' -f 2}
     # OpenRA engine version in spec file
-#    enpv=$(grep "^%define engine" < "$HOME"/OBS/home:fusion809/openra-ca/openra-ca.spec | cut -d ' ' -f 3)
-	mastn=$(comno)
-	specn=$(grep "VERSION" < $HOME/.local/share/openra-ca | cut -d ' ' -f 2)
-	comm=$(loge)
-	specm=$(grep "COMMIT" < $HOME/.local/share/openra-ca | cut -d ' ' -f 2)
+#    enpv=${grep "^%define engine" < "$HOME"/OBS/home:fusion809/openra-ca/openra-ca.spec | cut -d ' ' -f 3}
+	if ! [[ -f $HOME/.local/share/openra-ca ]]; then
+		printf "$HOME/.local/share/openra-ca does not exist, so creating it with touch.\n"
+		touch $HOME/.local/share/openra-ca || {printf "Touching it failed.\n" && return}
+	fi
+	mastn=${comno}
+	specn=${grep "VERSION" < $HOME/.local/share/openra-ca | cut -d ' ' -f 2}
+	comm=${loge}
+	specm=${grep "COMMIT" < $HOME/.local/share/openra-ca | cut -d ' ' -f 2}
+	modname=${grep "^PACKAGING_INSTALLER_NAME" < mod.config | cut -d '"' -f 2}
 
 	if [[ $specn == $mastn ]] && [[ $comm == $specm ]]; then
 		printf "OpenRA Combined Arms is up-to-date!\n"
 	else
 		printf "Cleaning repository.\n"
-		make clean || printf "Make cleanin' $GHUBO/CAmod failed.\n" && exit
+		make clean || {printf "Make cleanin' $GHUBO/CAmod failed.\n" && return}
 		# Setting version
 		printf "Setting version in $HOME/.local/share/openra-ca.\n"
 		echo "VERSION ${mastn}\nCOMMIT ${comm}" > $HOME/.local/share/openra-ca
-		make version VERSION=${mastn} || printf "Make versionin' $GHUBO/CAmod failed.\n" && exit
+		make version VERSION=${mastn} || {printf "Make versionin' $GHUBO/CAmod failed.\n" && return}
 		# Building $GHUBO/CAmod
 		printf "Building CAmod.\n"
-		make || printf "Making $GHUBO/CAmod failed.\n" && exit
+		make || {printf "Making $GHUBO/CAmod failed.\n" && return}
 		# Build AppImage
-		pushd packaging/linux || printf "pushdin' into packaging/linux.\n" && exit
-		./buildpackage.sh ${mastn} $HOME/Applications || printf "Building the AppImage with buildpackage.sh failed.\n" && exit
-		popd || printf "popdin' out of packaging/linux.\n" && exit
+		pushd packaging/linux || {printf "pushdin' into packaging/linux.\n" && return}
+		./buildpackage.sh ${mastn} . || {printf "Building AppImage failed.\n" && return}
+		if ls $HOME/Applications | grep "${modname}-" | grep AppImage > /dev/null 2>&1 ; then
+ 			printf "An existing AppImage seems to exist in $HOME/Applications, so deleting it, so we can replace it with the successfully build AppImage for this new version.\n"
+			rm $HOME/Applications/*${modname}-*.AppImage || {printf "Removing this AppImage failed.\n" && return}
+		fi
+		printf "Moving AppImage to $HOME/Applications.\n"
+		mv ${modname}-${mastn}.AppImage $HOME/Applications || {printf "Moving new AppImage to $HOME/Applications failed.\n" && return}
+		popd || {printf "popdin' out of packaging/linux.\n" && return}
 	fi
 
-	popd || printf "popdin' out of $GHUBO/CAmod.\n" && exit
+	popd || {printf "popdin' out of $GHUBO/CAmod.\n" && return}
 }
